03021402.09.41

+++ src/utility/coreMethod/address.ts
@@ -2,6 +2,8 @@ import {Address} from "../../mvc/model/address";
 import {AddressAddVM, AddressDeleteVM, AddressUpdateVM} from "../type/address";
 import {City} from "../../mvc/model/city";
 import mongoose from "mongoose";
+import {idIsNotValid} from "../validator";
+import {CareerHistory} from "../../mvc/model/careerHistory";
 
 export async function getCountOfAddress()
 {
@@ -22,7 +24,15 @@ export async function getAllAddress()
         .populate(
             {
                 path: 'city',
-                //populate: 'name'
+                select: 'title',
+                populate: {
+                    path: 'state',
+                    select: 'title',
+                    populate: {
+                        path: 'country',
+                        select: 'title',
+                    }
+                }
             }
         )
         .sort(
@@ -47,7 +57,15 @@ export async function getAddressById(id: string)
         .populate(
             {
                 path: 'city',
-                //populate: 'name'
+                select: 'title',
+                populate: {
+                    path: 'state',
+                    select: 'title',
+                    populate: {
+                        path: 'country',
+                        select: 'title',
+                    }
+                }
             }
         )
         .sort(
@@ -66,8 +84,90 @@ export async function getAddressById(id: string)
     }
 }
 
+export async function getAddressByIdAndFilter(id: string, filter: any)
+{
+    let currentAddress: any
+    if (filter)
+    {
+        currentAddress = await Address.findById(id)
+            .populate(
+                {
+                    path: 'city',
+                    select: 'title',
+                    populate: {
+                        path: 'state',
+                        select: 'title',
+                        populate: {
+                            path: 'country',
+                            select: 'title',
+                        }
+                    }
+                }
+            )
+            .select(`${filter}`)
+            .sort(
+                {
+                    'createDate': -1
+                }
+            );
+    }
+    else
+    {
+        currentAddress = await getAddressById(id)
+    }
+
+    if (currentAddress)
+    {
+        return currentAddress
+    }
+    else
+    {
+        return null
+    }
+}
+
+export async function getAddressByFilter(filter: any)
+{
+    let currentAddress = await Address.find()
+        .populate(
+            {
+                path: 'city',
+                select: 'title',
+                populate: {
+                    path: 'state',
+                    select: 'title',
+                    populate: {
+                        path: 'country',
+                        select: 'title',
+                    }
+                }
+            }
+        )
+        .select(`${filter}`)
+        .sort(
+            {
+                'createDate': -1
+            }
+        )
+
+    if (currentAddress)
+    {
+        return currentAddress
+    }
+    else
+    {
+        return null
+    }
+}
+
 export async function addNewAddress(entity: AddressAddVM): Promise<null | boolean>
 {
+    let currentAddressExist = await checkIfAddressWithTheSamePropertiesExist(entity)
+    if (currentAddressExist)
+    {
+        return true
+    }
+
     let currentCity = await City.findById(entity.city)
     if (!currentCity)
     {
@@ -102,7 +202,13 @@ export async function addNewAddress(entity: AddressAddVM): Promise<null | boolea
 
 export async function updateExistAddress(entity: AddressUpdateVM)
 {
-    if (!mongoose.isValidObjectId(entity.id))
+    let currentAddressExist = await checkIfAddressWithTheSamePropertiesExist(entity)
+    if (currentAddressExist)
+    {
+        return true
+    }
+
+    if (idIsNotValid(entity.id))
     {
         return null
     }
@@ -124,23 +230,21 @@ export async function updateExistAddress(entity: AddressUpdateVM)
 
 export async function deleteExistAddress(entity: AddressDeleteVM)
 {
-    await Address.findByIdAndRemove(entity.id)
-        .then(value =>
-        {
-            if (value)
-            {
-                console.log(value)
-                return true
-            }
-            else
-            {
-                console.log('Error while removing address!')
-                return false
-            }
-        })
-        .catch(reason =>
-        {
-            console.log(reason)
-            return false
-        })
+    if (idIsNotValid(entity.id))
+    {
+        return null
+    }
+    let result = await Address.findByIdAndRemove(entity.id)
+    return !!result;
+}
+
+
+async function checkIfAddressWithTheSamePropertiesExist(entity: AddressAddVM)
+{
+    let currentAddress = await Address.findOne({
+        city: entity.city,
+        restOfAddress: entity.restOfAddress
+    })
+    return !!currentAddress;
+
 }
